fixed small bugs

+++ src/filesystem.rs
@@ -170,7 +170,7 @@ pub fn read_image(name: [u8; 20]) -> Vec {
     let mut contents = Vec::new();
 
     for i in 0..file_size {
-        let byte = alloc::read_byte(file_start + i * 8) as u8;
+        let byte = alloc::read_byte(file_start + i * 8);
         contents.add(byte);
     }
